requirements.txt
pillow>=2.3
RPi.GPIO>=0.5
